src/test/kotlin/xyz/justinhorton/aoc2020/solutions/Day7Test.kt
package xyz.justinhorton.aoc2020.solutions

import org.junit.jupiter.api.BeforeEach
import org.junit.jupiter.api.Test
import kotlin.test.assertEquals

class Day7Test {
    private lateinit var day: Day7

    @BeforeEach
    fun beforeEach() {
        val inFile = Day7Test::class.java.getResource("input-day7.txt")
        day = Day7(inFile.readText())
    }

    @Test
    fun `test part 1 sample`() {
        assertEquals(4, Day7(SAMPLE).part1Solution.computeAnswer())
    }

    @Test
    fun `test part 1`() {
        assertEquals(252, day.part1Solution.computeAnswer())
    }

    @Test
    fun `test part 2 sample`() {
        assertEquals(126, Day7(SAMPLE2).part2Solution.computeAnswer())
    }

    @Test
    fun `test part 2`() {
        assertEquals(35487, day.part2Solution.computeAnswer())
    }
}

private val SAMPLE =
    """
    light red bags contain 1 bright white bag, 2 muted yellow bags.
    dark orange bags contain 3 bright white bags, 4 muted yellow bags.
    bright white bags contain 1 shiny gold bag.
    muted yellow bags contain 2 shiny gold bags, 9 faded blue bags.
    shiny gold bags contain 1 dark olive bag, 2 vibrant plum bags.
    dark olive bags contain 3 faded blue bags, 4 dotted black bags.
    vibrant plum bags contain 5 faded blue bags, 6 dotted black bags.
    faded blue bags contain no other bags.
    dotted black bags contain no other bags.
    """.trimIndent()

private val SAMPLE2 =
    """
    shiny gold bags contain 2 dark red bags.
    dark red bags contain 2 dark orange bags.
    dark orange bags contain 2 dark yellow bags.
    dark yellow bags contain 2 dark green bags.
    dark green bags contain 2 dark blue bags.
    dark blue bags contain 2 dark violet bags.
    dark violet bags contain no other bags.
    """.trimIndent()
